% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureCorrectMissingValue.R
\name{ensureCorrectMissingValue}
\alias{ensureCorrectMissingValue}
\title{The function ensure the missing value is correctly specified}
\usage{
ensureCorrectMissingValue(data, valueVar = "Value",
  flagObservationStatusVar = "flagObservationStatus",
  missingObservationFlag = "M", returnData = TRUE, getInvalidData = FALSE)
}
\arguments{
\item{data}{The data to be saved back to the database.}

\item{valueVar}{The variable to be tested.}

\item{flagObservationStatusVar}{The flag associated with the value.}

\item{missingObservationFlag}{The value of the observation status flag which
represents missing value.}

\item{returnData}{logical, whether the data should be returned}

\item{getInvalidData}{logical, this will skip the test and extract the data
that is invalid.}
}
\value{
If getInvalidData is FALSE, then the data is returned when the test
    is cleared, otherwise an error. If getInvalidData is TRUE, then the
    subset of the data that is invalid is returned.
}
\description{
The function ensure the missing value is correctly specified
}

