% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureCorrectTransfer.R
\name{ensureCorrectTransfer}
\alias{ensureCorrectTransfer}
\title{The function checks whether the transfer from parent to child commodity is
perform correctly.}
\usage{
ensureCorrectTransfer(parentData, childData, mappingTable, returnData = TRUE,
  getInvalidData = FALSE)
}
\arguments{
\item{parentData}{The animal slaughtered data from animal commodity.}

\item{childData}{The animal slaughtered data from meat commodity.}

\item{mappingTable}{The mapping table between the parent and the child.}

\item{returnData}{logical, whether the data should be returned}

\item{getInvalidData}{logical, this will skip the test and extract the data
that is invalid.}

\item{parentToChild}{logical, if true, slaughtered animal are transferred
from animal commodity to meat, otherwise the otherway around.}
}
\value{
If getInvalidData is FALSE, then the data is returned when the test
    is cleared, otherwise an error. If getInvalidData is TRUE, then the
    subset of the data that is invalid is returned.
}
\description{
For more details of a transfer, see function
\code{faoswsProduction:transferParentToChild},
}

