% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureDataInput.R
\name{ensureDataInput}
\alias{ensureDataInput}
\title{Function to ensure the data input is correct.}
\usage{
ensureDataInput(data, requiredColumn = NULL, returnData = TRUE)
}
\arguments{
\item{data}{The data to be checked.}

\item{requiredColumn}{The required column names the data must contain.}

\item{returnData}{logical, whether the data should be returned for pipe.}
}
\value{
If the data passes the check, the original data is returned,
    otherwise an error.
}
\description{
Function to ensure the data input is correct.
}

