% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureIdentityCalculated.R
\name{ensureIdentityCalculated}
\alias{ensureIdentityCalculated}
\title{Function to check whether the triplet (area harvested/production/yield) are
calculated whereever possible.}
\usage{
ensureIdentityCalculated(data, areaVar, yieldVar, prodVar, returnData = TRUE,
  normalised = TRUE, denormalisedKey = "measuredElement",
  getInvalidData = FALSE)
}
\arguments{
\item{data}{The data.table to be saved back to the SWS.}

\item{areaVar}{The column name corresponding to the area harvested.}

\item{yieldVar}{The column name corresponding to the yield.}

\item{prodVar}{The column name corresponding to produciton.}

\item{returnData}{logical, whether the data should be returned.}

\item{normalised}{logical, whether the data is normalised}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}

\item{getInvalidData}{logical, this will skip the test and extract the data
that is invalid.}
}
\value{
If getInvalidData is FALSE, then the data is returned when the test
    is cleared, otherwise an error. If getInvalidData is TRUE, then the
    subset of the data that is invalid is returned.
}
\description{
Function to check whether the triplet (area harvested/production/yield) are
calculated whereever possible.
}

