% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureNoConflictingZero.R
\name{ensureNoConflictingZero}
\alias{ensureNoConflictingZero}
\title{Function to ensure that two value column does not contain conflicting zero
values.}
\usage{
ensureNoConflictingZero(data, valueColumn1, valueColumn2, returnData = TRUE,
  normalised = TRUE, denormalisedKey = "measuredElement")
}
\arguments{
\item{data}{The data to be checked.}

\item{valueColumn1}{The first variable for comparison.}

\item{valueColumn2}{The second variable for comparison.}

\item{returnData}{logical, whether the data should be returned for pipe.}

\item{normalised}{logical, whether the data is normalised.}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}
}
\value{
If the data passes the check, the original data is returned,
    otherwise an error.
}
\description{
In the production domain, when production is zero, area harvested can not be
zero by definition and vice versa.
}

