% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureProtectedData.R
\name{ensureProtectedData}
\alias{ensureProtectedData}
\title{Function to check whether the data will over write protected data.}
\usage{
ensureProtectedData(data, domain = "agriculture", dataset = "aproduction",
  areaVar = "geographicAreaM49", itemVar = "measuredItemCPC",
  elementVar = "measuredElement", yearVar = "timePointYears",
  flagObservationVar = "flagObservationStatus",
  flagMethodVar = "flagMethod", flagTable = flagValidTable,
  returnData = TRUE, normalised = TRUE,
  denormalisedKey = "measuredElement")
}
\arguments{
\item{data}{The data.table object containing data to be saved back to the
database. The current implementation only accepts data in the normalised
form.}

\item{domain}{The domain name in the SWS where the data will be saved.}

\item{dataset}{The dataset name in the SWS where the data will be saved.}

\item{areaVar}{The column name corresponding to the geographic area.}

\item{itemVar}{The column name corresponding to the commodity item.}

\item{elementVar}{The column name corresponding to the measured element.}

\item{yearVar}{The column name corresponding to the year.}

\item{flagObservationVar}{The column name corresponding to the observation
status flag.}

\item{flagMethodVar}{The column name corresponding to the method flag.}

\item{flagTable}{The flag table containing the validity of the flags. see
data{faoswsFlag::flagValidTable} for an example.}

\item{returnData}{logical, whether the data should be returned.}

\item{normalised}{logical, whether the data is normalised}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}
}
\value{
If the data set passes the test, then the original data
    will be returned. Otherwise an error will be raised.
}
\description{
Certain data in the Statistical Working System should not be over
written by algorithms. Namely, official and semi-official
values. This function takes the data to be saved in the arguement
and pulls the exact same set of data from the data base, then
checks whether the matching set contains official or semi official
values.
}

