% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureProtectedDataFBS.R
\name{ensureProtectedDataFBS}
\alias{ensureProtectedDataFBS}
\title{Function to check whether the data will over write protected data.}
\usage{
ensureProtectedDataFBS(data, domain = "suafbs", dataset = "sua_unbalanced",
  areaVar = "geographicAreaM49", itemVar = "measuredItemChildCPC",
  elementVar = "measuredElement", yearVar = "timePointYears",
  flagObservationVar = "flagObservationStatus",
  flagMethodVar = "flagMethod", flagTable = flagValidTable,
  returnData = TRUE, normalised = TRUE,
  denormalisedKey = "measuredElement", getInvalidData = FALSE)
}
\arguments{
\item{data}{The data.table object containing data to be saved back to the
database. The current implementation only accepts data in the normalised
form.}

\item{domain}{The domain name in the SWS where the data will be saved.}

\item{dataset}{The dataset name in the SWS where the data will be saved.}

\item{areaVar}{The column name corresponding to the geographic area.}

\item{itemVar}{The column name corresponding to the commodity item.}

\item{elementVar}{The column name corresponding to the measured element.}

\item{yearVar}{The column name corresponding to the year.}

\item{flagObservationVar}{The column name corresponding to the observation
status flag.}

\item{flagMethodVar}{The column name corresponding to the method flag.}

\item{flagTable}{The flag table containing the validity of the flags. see
data{faoswsFlag::flagValidTable} for an example.}

\item{returnData}{logical, whether the data should be returned.}

\item{normalised}{logical, whether the data is normalised}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}

\item{getInvalidData}{logical, this will skip the test and extract the data
that is invalid.}
}
\value{
If getInvalidData is FALSE, then the data is returned when the test
    is cleared, otherwise an error. If getInvalidData is TRUE, then the
    subset of the data that is invalid is returned.
}
\description{
Certain data in the Statistical Working System should not be over
written by algorithms. Namely, official and semi-official
values. This function takes the data to be saved in the arguement
and pulls the exact same set of data from the data base, then
checks whether the matching set contains official or semi official
values.
}
\details{
DUPLICATION FOR FBS PURPOSES (Used in faoswsStandardization plugin)
}

