% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureTimeSeriesImputed.R
\name{ensureTimeSeriesImputed}
\alias{ensureTimeSeriesImputed}
\title{This function checks whether all time series within a dataset are imputed.}
\usage{
ensureTimeSeriesImputed(data, key, valueColumn = "Value",
  observationFlagVar = "flagObservationStatus", returnData = TRUE,
  normalised = TRUE, denormalisedKey = "measuredElement",
  getInvalidData = FALSE)
}
\arguments{
\item{data}{The data to be saved back to the database}

\item{key}{The key which splits the data into individual timeseries.
Generally the set c('geographicAreaM49', 'measuredItemCPC',
'measuredElement').}

\item{valueColumn}{The column which contains the numeric values.}

\item{observationFlagVar}{The variable name which corresponds to the
observation status flag.}

\item{returnData}{logical, whether the data should be returned}

\item{normalised}{logical, whether the data is normalised}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}

\item{getInvalidData}{logical, this will skip the test and extract the data
that is invalid.}
}
\value{
If getInvalidData is FALSE, then the data is returned when the test
    is cleared, otherwise an error. If getInvalidData is TRUE, then the
    subset of the data that is invalid is returned.
}
\description{
This function checks whether all time series within a dataset are imputed.
}

