% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureValueRange.R
\name{ensureValueRange}
\alias{ensureValueRange}
\title{This function checks whether the values are within valid range}
\usage{
ensureValueRange(data, ensureColumn, min = 0, max = Inf,
  includeEndPoint = TRUE, returnData = TRUE)
}
\arguments{
\item{data}{The data to be saved back to the database}

\item{ensureColumn}{The variable to be ensured.}

\item{min}{The minimum bound of the variable}

\item{max}{The maximum bound of the variable}

\item{includeEndPoint}{Whether the end point of minimum and maximum should
be included.}

\item{returnData}{logical, whether the data should be returned}
}
\value{
The same data if all time series are imputed, otherwise an error.
}
\description{
This function checks whether the values are within valid range
}

