% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoswsEnsure-package.R
\docType{package}
\name{faoswsEnsure-package}
\alias{faoswsEnsure}
\alias{faoswsEnsure-package}
\title{The package host standard check function for the Statistical Working
    System (SWS).}
\description{
Package to ensure data and input quality
}
\author{
Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

