% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureValueRange.R
\name{ensureValueRange}
\alias{ensureValueRange}
\title{This function checks whether the values are within valid range}
\usage{
ensureValueRange(data, ensureColumn, min = 0, max = Inf,
  includeEndPoint = TRUE, returnData = TRUE, getInvalidData = FALSE)
}
\arguments{
\item{data}{The data to be saved back to the database}

\item{ensureColumn}{The variable to be ensured.}

\item{min}{The minimum bound of the variable}

\item{max}{The maximum bound of the variable}

\item{includeEndPoint}{Whether the end point of minimum and maximum should
be included.}

\item{returnData}{logical, whether the data should be returned}

\item{getInvalidData}{logical, this will skip the test and extract the data
that is invalid.}
}
\value{
If getInvalidData is FALSE, then the data is returned when the test
    is cleared, otherwise an error. If getInvalidData is TRUE, then the
    subset of the data that is invalid is returned.
}
\description{
This function checks whether the values are within valid range
}

